@@ -10,7 +10,7 @@
     ],
     "minimum-stability": "dev",
     "require": {
-      "php": "^7.3|^8.0",
+      "php": "^7.4|^8.0",
       "illuminate/support": "^5.5|^6.0|^7.0|^8.0|^9.0"
     },
     "require-dev": {
